requirements.txt
fastapi[standard]
SQLAlchemy>=2.0.40
pydantic>=2.11.3
python-decouple>=3.8
psycopg2>=2.9.10
alembic>=1.15.2